requirement.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.28
urllib3==2.2.1
wheel==0.43.0
